requirements.txt
argon2-cffi==21.3.0
Flask==2.2.5
Flask-Session==0.4.1
joblib==1.2.0
numpy==1.24.3
pandas==2.0.1
SQLAlchemy==2.0.12
keras==2.13.1
scikit-learn==1.2.1
tensorflow==2.13.0
PyMySQL==1.0.3
music21==9.1.0
midi2audio==0.1.1
boto3==1.28.62
cryptography==39.0.1
pytorch
pypianoroll
xmldataset
ipdb
matplotlib